+++ __tests__/__system__/cli/revert/__scripts__/invalid_credmgr.sh
@@ -1,0 +1,10 @@
+#!/bin/bash
+
+credMgr=`zowe config get CredentialManager`
+zowe config reset CredentialManager
+
+zowe scs revert -f
+
+zowe config set CredentialManager $credMgr
+
+exit $?