+++ __tests__/__system__/cli/revert/__scripts__/revert_help.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+zowe scs revert --help
+
+zowe scs revert -h --rfj
+
+exit $?