+++ __tests__/__system__/cli/update_revert/__scripts__/revert_success.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+zowe scs revert -f
+
+exit $?